@@ -1,8 +1,6 @@
 module github.com/cerbos/cerbos/hack/tools/testsplit
 
-go 1.23
-
-toolchain go1.24.0
+go 1.24
 
 require (
 	github.com/alecthomas/kong v1.8.1
