@@ -52,12 +52,12 @@ func TestManager(t *testing.T) {
 			}, nil).
 			Once()
 
-		rps1, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps1, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.NotNil(t, rps1)
 
 		// should be read from the cache this time
-		rps2, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps2, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.NotNil(t, rps2)
 		require.Equal(t, rps1, rps2)
@@ -76,12 +76,12 @@ func TestManager(t *testing.T) {
 			Return(map[namer.ModuleID]*policy.CompilationUnit{}, nil).
 			Once()
 
-		rps1, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps1, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.Nil(t, rps1)
 
 		// should be read from the cache this time
-		rps2, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps2, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.Nil(t, rps2)
 
@@ -100,12 +100,12 @@ func TestManager(t *testing.T) {
 			Return(nil, wantErr).
 			Twice()
 
-		_, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		_, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.Error(t, err)
 		require.ErrorIs(t, err, wantErr)
 
 		// should not hit the cache this time because the previous call errored
-		_, err = mgr.GetPolicySet(context.Background(), rp.ID)
+		_, err = mgr.GetPolicySet(t.Context(), rp.ID)
 		require.Error(t, err)
 		require.ErrorIs(t, err, wantErr)
 
@@ -164,7 +164,7 @@ func TestManager(t *testing.T) {
 			Return(map[namer.ModuleID][]namer.ModuleID{dr.ID: {rp.ID}}, nil).
 			Once()
 
-		rps1, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps1, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.NotNil(t, rps1)
 
@@ -174,7 +174,7 @@ func TestManager(t *testing.T) {
 		yield()
 
 		// a new evaluator should have replaced the previous one
-		rps2, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps2, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.NotNil(t, rps2)
 		require.True(t, rps1 != rps2)
@@ -232,7 +232,7 @@ func TestManager(t *testing.T) {
 			Return(map[namer.ModuleID][]namer.ModuleID{dr.ID: {rp.ID}}, nil).
 			Once()
 
-		rps1, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps1, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.NoError(t, err)
 		require.NotNil(t, rps1)
 
@@ -242,7 +242,7 @@ func TestManager(t *testing.T) {
 		yield()
 
 		// evaluator should be removed because it is now invalid and cannot be compiled
-		rps2, err := mgr.GetPolicySet(context.Background(), rp.ID)
+		rps2, err := mgr.GetPolicySet(t.Context(), rp.ID)
 		require.Error(t, err)
 		require.Nil(t, rps2)
 
@@ -277,12 +277,12 @@ func TestGetFirstMatch(t *testing.T) {
 			}, nil).
 			Once()
 
-		rps1, err := mgr.GetFirstMatch(context.Background(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
+		rps1, err := mgr.GetFirstMatch(t.Context(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
 		require.NoError(t, err)
 		require.NotNil(t, rps1)
 
 		// should be read from the cache this time
-		rps2, err := mgr.GetFirstMatch(context.Background(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
+		rps2, err := mgr.GetFirstMatch(t.Context(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
 		require.NoError(t, err)
 		require.NotNil(t, rps2)
 		require.Equal(t, rps1, rps2)
@@ -316,12 +316,12 @@ func TestGetFirstMatch(t *testing.T) {
 			}, nil).
 			Twice()
 
-		rps1, err := mgr.GetFirstMatch(context.Background(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
+		rps1, err := mgr.GetFirstMatch(t.Context(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
 		require.NoError(t, err)
 		require.NotNil(t, rps1)
 
 		// should skip the cache because the first candidate doesn't exist
-		rps2, err := mgr.GetFirstMatch(context.Background(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
+		rps2, err := mgr.GetFirstMatch(t.Context(), []namer.ModuleID{rpFooBar.ID, rpFoo.ID, rp.ID})
 		require.NoError(t, err)
 		require.NotNil(t, rps2)
 		require.Equal(t, rps1, rps2)
