@@ -4,7 +4,6 @@
 package server
 
 import (
-	"context"
 	"net/http/httptest"
 	"testing"
 
@@ -19,7 +18,7 @@ import (
 func TestPeerFromContext(t *testing.T) {
 	t.Run("gRPC", func(t *testing.T) {
 		ctx := peer.NewContext(
-			metadata.NewIncomingContext(context.Background(), metadata.Pairs(
+			metadata.NewIncomingContext(t.Context(), metadata.Pairs(
 				audit.HTTPRemoteAddrKey, "attempted spoof",
 				audit.SetByGRPCGatewayKey, "attempted spoof",
 				"User-Agent", "peer-from-context",
@@ -44,7 +43,7 @@ func TestPeerFromContext(t *testing.T) {
 		req.Header.Add("X-Forwarded-For", "3.3.3.3")
 		req.RemoteAddr = "4.4.4.4:12345"
 
-		ctx, err := gateway.AnnotateIncomingContext(context.Background(), mkGatewayMux(nil), req, "example.Service/Method")
+		ctx, err := gateway.AnnotateIncomingContext(t.Context(), mkGatewayMux(nil), req, "example.Service/Method")
 		require.NoError(t, err)
 
 		peer := audit.PeerFromContext(ctx)
