@@ -51,12 +51,12 @@ type testParam struct {
 type testParamGen func(*testing.T) testParam
 
 func TestServer(t *testing.T) {
-	logging.InitLogging(context.Background(), "ERROR")
+	logging.InitLogging(t.Context(), "ERROR")
 
 	t.Run("store=disk", func(t *testing.T) {
 		tpg := func(t *testing.T) testParam {
 			t.Helper()
-			ctx, cancelFunc := context.WithCancel(context.Background())
+			ctx, cancelFunc := context.WithCancel(t.Context())
 			t.Cleanup(cancelFunc)
 
 			dir := test.PathToDir(t, "store")
@@ -81,7 +81,7 @@ func TestServer(t *testing.T) {
 		tpg := func(version bundle.Version) func(t *testing.T) testParam {
 			return func(t *testing.T) testParam {
 				t.Helper()
-				ctx, cancelFunc := context.WithCancel(context.Background())
+				ctx, cancelFunc := context.WithCancel(t.Context())
 				t.Cleanup(cancelFunc)
 
 				dir := test.PathToDir(t, filepath.Join("bundle", fmt.Sprintf("v%d", version)))
@@ -251,7 +251,7 @@ func TestAdminService(t *testing.T) {
 	tpg := func(t *testing.T) testParam {
 		t.Helper()
 
-		ctx, cancelFunc := context.WithCancel(context.Background())
+		ctx, cancelFunc := context.WithCancel(t.Context())
 		t.Cleanup(cancelFunc)
 
 		store, err := sqlite3.NewStore(ctx, &sqlite3.Conf{DSN: fmt.Sprintf("%s?_fk=true", filepath.Join(t.TempDir(), "cerbos.db"))})
@@ -311,7 +311,7 @@ func startServer(t *testing.T, conf *Conf, tpg testParamGen) {
 
 	tp := tpg(t)
 
-	ctx, cancelFunc := context.WithCancel(context.Background())
+	ctx, cancelFunc := context.WithCancel(t.Context())
 	t.Cleanup(cancelFunc)
 
 	auditLog := audit.NewNopLog()
