@@ -7,7 +7,6 @@
 package internal
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -28,30 +27,30 @@ func TestSuiteReloadable(store storage.Store, initFn, addFn, deleteFn MutateStor
 		if initFn != nil {
 			expectedLen = 1
 			require.NoError(t, initFn())
-			require.NoError(t, r.Reload(context.Background()))
+			require.NoError(t, r.Reload(t.Context()))
 		}
 
-		policies, err := store.ListPolicyIDs(context.Background(), storage.ListPolicyIDsParams{})
+		policies, err := store.ListPolicyIDs(t.Context(), storage.ListPolicyIDsParams{})
 		require.NoError(t, err)
 		require.Len(t, policies, expectedLen)
 
 		require.NoError(t, addFn())
 
-		policies, err = store.ListPolicyIDs(context.Background(), storage.ListPolicyIDsParams{})
+		policies, err = store.ListPolicyIDs(t.Context(), storage.ListPolicyIDsParams{})
 		require.NoError(t, err)
 		require.Len(t, policies, expectedLen)
 
-		require.NoError(t, r.Reload(context.Background()))
+		require.NoError(t, r.Reload(t.Context()))
 
-		policies, err = store.ListPolicyIDs(context.Background(), storage.ListPolicyIDsParams{})
+		policies, err = store.ListPolicyIDs(t.Context(), storage.ListPolicyIDsParams{})
 		require.NoError(t, err)
 		require.Greater(t, len(policies), expectedLen)
 
 		require.NoError(t, deleteFn())
 
-		require.NoError(t, r.Reload(context.Background()))
+		require.NoError(t, r.Reload(t.Context()))
 
-		policies, err = store.ListPolicyIDs(context.Background(), storage.ListPolicyIDsParams{})
+		policies, err = store.ListPolicyIDs(t.Context(), storage.ListPolicyIDsParams{})
 		require.NoError(t, err)
 		require.Len(t, policies, expectedLen)
 	}
