@@ -120,7 +120,7 @@ func runPolicyTests(t *testing.T, eng *engine.Engine, tc *TestCase) (*policyv1.T
 
 	config := tc.GetConfig()
 
-	return Verify(context.Background(), os.DirFS(dir), eng, Config{
+	return Verify(t.Context(), os.DirFS(dir), eng, Config{
 		ExcludedResourcePolicyFQNs:  util.ToStringSet(config.GetExcludedResourcePolicyFqns()),
 		ExcludedPrincipalPolicyFQNs: util.ToStringSet(config.GetExcludedPrincipalPolicyFqns()),
 		IncludedTestNamesRegexp:     config.GetIncludedTestNamesRegexp(),
@@ -293,7 +293,7 @@ func Test_doVerify(t *testing.T) {
 				}
 				table := genTable(t, optionResources != external, optionPrincipals != external)
 				fsys["leave_request_test.yaml"] = newMapFile(table)
-				result, err := Verify(context.Background(), fsys, eng, Config{})
+				result, err := Verify(t.Context(), fsys, eng, Config{})
 				is := require.New(t)
 				is.NoError(err)
 				is.Len(result.Suites, 1)
@@ -309,7 +309,7 @@ func Test_doVerify(t *testing.T) {
 
 		table := genTable(t, false, false)
 		fsys["leave_request_test.yaml"] = newMapFile(table)
-		result, err := Verify(context.Background(), fsys, eng, Config{})
+		result, err := Verify(t.Context(), fsys, eng, Config{})
 		is := require.New(t)
 		is.NoError(err)
 		is.Len(result.Suites, 1)
@@ -323,7 +323,7 @@ func Test_doVerify(t *testing.T) {
 
 		table := genTable(t, false, false)
 		fsys["leave_request_test.yaml"] = newMapFile(table)
-		result, err := Verify(context.Background(), fsys, eng, Config{})
+		result, err := Verify(t.Context(), fsys, eng, Config{})
 		is := require.New(t)
 		is.NoError(err)
 		is.Len(result.Suites, 1)
@@ -340,7 +340,7 @@ func Test_doVerify(t *testing.T) {
 			fsys[dir+"/leave_request_test.yaml"] = newMapFile(ts)
 		}
 
-		result, err := Verify(context.Background(), fsys, eng, Config{})
+		result, err := Verify(t.Context(), fsys, eng, Config{})
 		is := require.New(t)
 		is.NoError(err)
 		is.Len(result.Suites, 3)
@@ -358,7 +358,7 @@ func Test_doVerify(t *testing.T) {
 		fsys[filepath.Join(util.TestDataDirectory, principalsFileName)+".yaml"] = newMapFile(principals)
 		fsys["leave_request_test.yaml"] = newMapFile(ts)
 
-		result, err := Verify(context.Background(), fsys, eng, Config{})
+		result, err := Verify(t.Context(), fsys, eng, Config{})
 		is := require.New(t)
 		is.NoError(err)
 		is.Len(result.Suites, 1)
@@ -372,7 +372,7 @@ func mkEngine(t *testing.T) *engine.Engine {
 
 	dir := test.PathToDir(t, "store")
 
-	ctx, cancelFunc := context.WithCancel(context.Background())
+	ctx, cancelFunc := context.WithCancel(t.Context())
 	t.Cleanup(cancelFunc)
 
 	store, err := disk.NewStore(ctx, &disk.Conf{Directory: dir})
