@@ -22,7 +22,7 @@ func TestServe(t *testing.T) {
 	// run twice to make sure that global state initialization is idempotent
 	for run := 0; run < 2; run++ {
 		t.Run(fmt.Sprintf("run_%d", run), func(t *testing.T) {
-			ctx, cancel := context.WithCancel(context.Background())
+			ctx, cancel := context.WithCancel(t.Context())
 			t.Cleanup(cancel)
 
 			grpcListenAddr, err := util.GetFreeListenAddr()
