@@ -414,7 +414,7 @@ func callInStringStringOutBool(fn func(string, string) (bool, error)) functions.
 
 		retVal, err := fn(string(lhs), string(rhs))
 		if err != nil {
-			return types.NewErr(err.Error()) //nolint:govet
+			return types.NewErr("%s", err.Error()) //nolint:govet
 		}
 
 		return types.DefaultTypeAdapter.NativeToValue(retVal)
